refactor: next config to match next next-pwa implementation, and add caching on every stylesheet element on client, and extend runtimeCaching to images
