+++ next.config.mjs
@@ -1,0 +1,50 @@
+import withPWAInit from "@ducanh2912/next-pwa";
+
+const isDev = process.env.NODE_ENV !== "production";
+
+const withPWA = withPWAInit({
+  cacheOnFrontEndNav: true,
+  aggressiveFrontEndNavCaching: true,
+  reloadOnOnline: true,
+  dest: "public",
+  customWorkerSrc: "./workers",
+  exclude: [
+    // add buildExcludes here
+    ({ asset, compilation }) => {
+      if (
+        asset.name.startsWith("server/") ||
+        asset.name.match(/^((app-|^)build-manifest\.json|react-loadable-manifest\.json)$/) ||
+        asset.name.match(/^(image|video|audio|font)\/.*$/i)
+      ) {
+        return true;
+      }
+      if (isDev && !asset.name.startsWith("static/runtime/")) {
+        return true;
+      }
+      return false;
+    }
+  ],
+  extendDefaultRuntimeCaching: true,
+  workboxOptions: {
+    exclude: [/app-build-manifest\.json$/,],
+    runtimeCaching: [
+      {
+        urlPattern: /^https:\/\/.*\.(png|jpg|jpeg|svg|gif|webp)/i,
+        handler: 'CacheFirst',
+        options: {
+          cacheName: 'image-cache',
+          expiration: {
+            maxEntries: 50,
+            maxAgeSeconds: 30 * 24 * 60 * 60 // 30 days
+          }
+        }
+      }
+    ]
+  }
+});
+
+export default withPWA({
+  experimental: {
+    optimizeCss: true,
+  },
+});
